refactor(storage): move retryability to StorageError + cover Db::open path
/polish phase 1+2 で 3 件指摘を一括適用:

1. **Codex P2**: `Db::open` 時の SQLite contention は `SaeError::Storage`
   (not Sync) で帰るため、当初の `Sync(e) if e.is_retryable()` arm を
   bypass していた → `Storage(e) if e.is_retryable()` arm を追加し
   `index`/`rebuild` 起動時の WAL contention も TempFailure ルート
2. **Quality HIGH**: T-260a〜T-260e は output.rs:385 の T-260 と衝突。
   a/b/c suffix は SUBTEST 専用なので T-353〜T-357 に renumber
3. **Quality medium + Reuse low**: `rusqlite::Error::SqliteFailure` の
   inline literal 4 site 重複 → `storage::sqlite_failure(code, ext)` test
   helper を `#[cfg(test)] pub(crate)` で抽出

アーキテクチャ変更:

- `StorageError::is_retryable()` を新設 (matrix logic を所有)
- `SyncError::is_retryable()` は `StorageError::is_retryable()` に
  delegate (Client variants は常に false)
- helper fn `is_retryable_sqlite_error` / `is_retryable_io_error` は
  `storage/types.rs` に集約
- `tools.rs::error_code` に 2 つの arm を `CantCreat` 既定の **前** に追加
  - `Self::Storage(e) if e.is_retryable() => TempFailure`
  - `Self::Sync(e) if e.is_retryable() => TempFailure`

テスト:
- T-353〜T-357 (sync.rs:tests): SyncError 経由の delegation pin
- T-388 (tools.rs:tests): Sync(Storage(BUSY)) → TEMP_FAIL
- T-389 (tools.rs:tests): Sync(Storage(Open)) → CANT_CREAT (arm 順序 pin)
- T-390 (tools.rs:tests, **新規**): direct Storage(BUSY) → TEMP_FAIL
  (Codex P2 finding を pin)

363 tests PASS / clippy clean。
